# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AYLink Agent 的 HTTP API 文档。",
        "title": "AYLink Agent API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/adb/pair": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "使用 Android 无线调试配对地址、端口和配对码完成 ADB 配对。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADB"
                ],
                "summary": "配对无线调试设备",
                "parameters": [
                    {
                        "description": "配对参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ADBPairRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ADBPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/adb/server/kill": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "关闭本机 ADB server。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADB"
                ],
                "summary": "关闭 ADB 服务",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/adb/server/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "启动本机 ADB server。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADB"
                ],
                "summary": "启动 ADB 服务",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/adb/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回 ADB 服务地址、已解析的 ADB 可执行文件和当前设备列表。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ADB"
                ],
                "summary": "获取 ADB 状态",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ADBStatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "adb.Device": {
            "type": "object",
            "properties": {
                "serial": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "adb.ResolvedBinary": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "handler.ADBPairRequest": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "pairingCode": {
                    "type": "string"
                },
                "pairingPort": {
                    "type": "integer"
                }
            }
        },
        "handler.ADBPairResponse": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "pairingPort": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handler.ADBStatusResponse": {
            "type": "object",
            "properties": {
                "binary": {
                    "$ref": "#/definitions/adb.ResolvedBinary"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/adb.Device"
                    }
                },
                "serverAddress": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "messageKey": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handler.ErrorPayload"
                }
            }
        },
...
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 Bearer 访问令牌，例如：Bearer {accessToken}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}